// src/app/(admin-panel)/inventories/auto-generated-stock/stock-adder.tsx
export type TStockRow = {
  barcode: string;
  colorId: string;
  sizeId: string;
  quantity: string;
  costPerItem: string;
  paidAmount: string;
  dueAmount: string;
  productDate: string;
  shelfLife: string;
  expireDate: string;
  supplierName: string;
};